+++ examples/diploid_fixed_sh_ind.cc
@@ -70,7 +70,7 @@ main(int argc, char **argv)
                                  + 0.667 * (theta_neutral + theta_del))));
             unsigned generation = 0;
             const auto mmodel = [&pop, &r, &generation, s, h,
-                                 pselected](std::queue<std::size_t> &recbin,
+                                 pselected](fwdpp::flagged_mutation_queue &recbin,
                                             singlepop_t::mcont_t &mutations) {
                 return fwdpp::infsites_popgenmut(
                     recbin, mutations, r.get(), pop.mut_lookup, generation,