+++ testsuite/unit/test_mutate_recombine.cc
@@ -23,8 +23,8 @@ BOOST_FIXTURE_TEST_CASE(test_boundary, slocuspop_objects)
     BOOST_REQUIRE_EQUAL(pop.mutations.back().pos, 1.5);
     std::vector<fwdpp::uint_t> new_mutations;
     std::vector<double> breakpoints{ 1.5, std::numeric_limits<double>::max() };
-    auto q = fwdpp::fwdpp_internal::make_mut_queue(pop.mcounts);
-    BOOST_REQUIRE_EQUAL(q.empty(), true);
+    auto q = fwdpp::make_gamete_queue(pop.gametes);
+    BOOST_REQUIRE_EQUAL(q.get().empty(), true);
 
     auto new_gamete_index = fwdpp::mutate_recombine(
         new_mutations, breakpoints, 0, 1, pop.gametes, pop.mutations, q,
@@ -77,8 +77,8 @@ BOOST_FIXTURE_TEST_CASE(test_boundary_with_recurrent_mutation,
     // The new mutation at 1.5 is at the end
     std::vector<fwdpp::uint_t> new_mutations(1, pop.mutations.size() - 1);
     std::vector<double> breakpoints{ 1.5, std::numeric_limits<double>::max() };
-    auto q = fwdpp::fwdpp_internal::make_mut_queue(pop.mcounts);
-    BOOST_REQUIRE_EQUAL(q.empty(), true);
+    auto q = fwdpp::make_gamete_queue(pop.gametes);
+    BOOST_REQUIRE_EQUAL(q.get().empty(), true);
 
     auto new_gamete_index = fwdpp::mutate_recombine(
         new_mutations, breakpoints, 0, 1, pop.gametes, pop.mutations, q,